Firefox 63.0 base

+++ mozilla-release/testing/raptor/raptor/control_server.py
@@ -87,6 +87,9 @@ class RaptorControlServer():
         self.port = None
         self.results_handler = results_handler
         self.browser_proc = None
+        self._finished = False
+        self.device = None
+        self.app_name = None
 
     def start(self):
         config_dir = os.path.join(here, 'tests')
@@ -111,18 +114,25 @@ class RaptorControlServer():
         self.server = httpd
 
     def shutdown_browser(self):
-        LOG.info("shutting down browser (pid: %d)" % self.browser_proc.pid)
+        if self.device is not None:
+            LOG.info("shutting down android app %s" % self.app_name)
+        else:
+            LOG.info("shutting down browser (pid: %d)" % self.browser_proc.pid)
         self.kill_thread = threading.Thread(target=self.wait_for_quit)
         self.kill_thread.daemon = True
         self.kill_thread.start()
 
-    def wait_for_quit(self, timeout=5):
+    def wait_for_quit(self, timeout=15):
         """Wait timeout seconds for the process to exit. If it hasn't
         exited by then, kill it.
         """
-        self.browser_proc.wait(timeout)
-        if self.browser_proc.poll() is None:
-            self.browser_proc.kill()
+        if self.device is not None:
+            self.device.stop_application(self.app_name)
+        else:
+            self.browser_proc.wait(timeout)
+            if self.browser_proc.poll() is None:
+                self.browser_proc.kill()
+        self._finished = True
 
     def stop(self):
         LOG.info("shutting down control server")
